# gestepm-rest-api/src/main/resources/apis/gestepm-core-api/api-spec.yaml
openapi: 3.0.3
info:
  title: GestEPM Core API
  version: 0.1.0
  description: GestEPM Core API
  contact:
    name: Carlos Devesa Pérez
servers:
  - url: 'http://localhost:8080'
paths:
...
  /v1/shares/no-programmed:
    get:
      summary: V1 - List no programmed shares
      responses:
        '200':
          $ref: '#/components/responses/NoProgrammedShareList'
      operationId: list-no-programmed-shares-v1
      description: list-no-programmed-shares-v1
      parameters:
        - $ref: ../../base-api-spec.yaml#/components/parameters/api-opt-meta
        - $ref: ../../base-api-spec.yaml#/components/parameters/api-opt-links
        - $ref: ../../base-api-spec.yaml#/components/parameters/api-opt-expand
        - $ref: ../../base-api-spec.yaml#/components/parameters/filter-pagination-offset
        - $ref: ../../base-api-spec.yaml#/components/parameters/filter-pagination-limit
        - $ref: ../../base-api-spec.yaml#/components/parameters/filter-order
        - $ref: ../../base-api-spec.yaml#/components/parameters/filter-order-by
        - schema:
            type: array
            items:
              type: integer
          in: query
          name: ids
          description: Filter by no programmed share identificator.
        - schema:
            type: array
            items:
              type: integer
          in: query
          name: userIds
          description: Filter by user identificator.
        - schema:
            type: array
            items:
              type: integer
          in: query
          name: projectIds
          description: Filter by project identificator.
        - schema:
            type: string
            format: date-time
          in: query
          name: startDate
          description: Filter by start date after.
        - schema:
            type: string
            format: date-time
          in: query
          name: endDate
          description: Filter by end date before.
        - schema:
            type: string
            enum:
              - NOT_FINISHED
              - FINISHED
          in: query
          name: status
          description: Filter by no programmed share status.
      tags:
        - no-programmed-share-v1
    post:
      summary: V1 - Create no programmed share
      operationId: create-no-programmed-share-v1
      responses:
        '200':
          $ref: '#/components/responses/NoProgrammedShare'
        '404':
          description: user-not-found | project-not-found
          content:
            application/json:
              schema:
                $ref: ../../base-api-spec.yaml#/components/schemas/ResError
      description: create-no-programmed-share-v1
      tags:
        - no-programmed-share-v1
      requestBody:
        $ref: '#/components/requestBodies/CreateNoProgrammedShare'
  '/v1/shares/no-programmed/{shareId}':
    parameters:
      - $ref: '#/components/parameters/Share-Id'
    get:
      summary: V1 - Find no programmed share
      tags:
        - no-programmed-share-v1
      responses:
        '200':
          $ref: '#/components/responses/NoProgrammedShare'
        '404':
          description: no-programmed-share-not-found
          content:
            application/json:
              schema:
                $ref: ../../base-api-spec.yaml#/components/schemas/ResError
      operationId: find-no-programmed-share-by-id-v1
      description: find-no-programmed-share-by-id-v1
      parameters:
        - $ref: ../../base-api-spec.yaml#/components/parameters/api-opt-meta
        - $ref: ../../base-api-spec.yaml#/components/parameters/api-opt-links
        - $ref: ../../base-api-spec.yaml#/components/parameters/api-opt-expand
        - $ref: ../../base-api-spec.yaml#/components/parameters/filter-locale
    patch:
      summary: V1 - Update no programmed share
      operationId: update-no-programmed-share-v1
      responses:
        '200':
          $ref: '#/components/responses/NoProgrammedShare'
        '404':
          description: no-programmed-share-not-found
          content:
            application/json:
              schema:
                $ref: ../../base-api-spec.yaml#/components/schemas/ResError
      description: update-no-programmed-share-v1
      tags:
        - no-programmed-share-v1
      requestBody:
        $ref: '#/components/requestBodies/UpdateNoProgrammedShare'
    delete:
      summary: V1 - Delete no programmed share
      operationId: delete-no-programmed-share-v1
      responses:
        '200':
          description: '**OK**'
          content:
            application/json:
              schema:
                $ref: ../../base-api-spec.yaml#/components/schemas/ResSuccess
        '404':
          description: no-programmed-share-not-found
          content:
            application/json:
              schema:
                $ref: ../../base-api-spec.yaml#/components/schemas/ResError
      description: delete-no-programmed-share-v1
      tags:
        - no-programmed-share-v1
...
components:
  schemas:
...
    Family:
      title: Family
      x-stoplight:
        id: wr1b8bilbahz0
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          x-stoplight:
            id: 0q1hdsc073x4b
    Inspection:
      title: Inspection
      x-stoplight:
        id: isocdt4oo7s7y
      type: object
      required:
        - id
        - share
        - action
        - startDate
        - firstTechnical
      properties:
        id:
          type: integer
        share:
          $ref: '#/components/schemas/NoProgrammedShare'
        action:
          type: string
          x-stoplight:
            id: 4v0eavn531ar2
          enum:
            - INTERVENTION
            - DIAGNOSIS
            - FOLLOWING
        startDate:
          type: string
          x-stoplight:
            id: 14nn1910i1mwc
          format: date-time
        endDate:
          type: string
          x-stoplight:
            id: 1blghe5381za3
          format: date-time
        description:
          type: string
          x-stoplight:
            id: b4w2w0bii3k0n
        firstTechnical:
          $ref: '#/components/schemas/User'
        secondTechnical:
          $ref: '#/components/schemas/User'
        signature:
          type: string
          x-stoplight:
            id: 8y6v5dl5jna1k
        operatorSignature:
          type: string
          x-stoplight:
            id: o4vgea6d0oj11
        clientName:
          type: string
          x-stoplight:
            id: y5zosfb602nzz
        materials:
          x-stoplight:
            id: 96cv7joi0oiwc
          type: array
          items:
            $ref: '#/components/schemas/Material'
        materialsFile:
          type: string
          x-stoplight:
            id: 8lfzk9pa7wyrg
        materialsFileName:
          type: string
          x-stoplight:
            id: 790fxpi66ikvh
        equipmentHours:
          type: integer
          x-stoplight:
            id: jlcvcmslrddth
        topicId:
          type: integer
          x-stoplight:
            id: jefh8p0m16cgw
        files:
          type: array
          x-stoplight:
            id: lxtp67qj9tr4z
          items:
            $ref: '#/components/schemas/ShareFile'
        updatedAt:
          type: string
          x-stoplight:
            id: ymzpdb9dd1an7
          format: date-time
        updatedBy:
          $ref: '#/components/schemas/User'
    NoProgrammedShare:
      title: NoProgrammedShare
      x-stoplight:
        id: nkbdyy6acj6uf
      type: object
      required:
        - id
        - user
        - project
        - startDate
        - description
        - family
        - subFamily
        - state
      properties:
        id:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        project:
          $ref: '#/components/schemas/Project'
        startDate:
          type: string
          x-stoplight:
            id: yizr915id0hco
          format: date-time
        endDate:
          type: string
          x-stoplight:
            id: hk6tamjfxzyw9
          format: date-time
        description:
          type: string
          x-stoplight:
            id: gdj67dr4ncyyd
        family:
          $ref: '#/components/schemas/Family'
        subFamily:
          $ref: '#/components/schemas/SubFamily'
        topicId:
          type: integer
          x-stoplight:
            id: j8jp9c1aphphk
        forumTitle:
          type: string
          x-stoplight:
            id: l7fu2mhea9mhk
        state:
          type: string
          x-stoplight:
            id: anff5xkbzmal1
          enum:
            - NEW
            - INITIALIZED
            - IN_PROGRESS
            - CLOSED
        lastDiagnosis:
          type: integer
          x-stoplight:
            id: g8rnuovvahzqf
        inspections:
          x-stoplight:
            id: vbc9pywk2s6rp
          type: array
          items:
            $ref: '#/components/schemas/Inspection'
        files:
          type: array
          x-stoplight:
            id: bm1bnujfonrek
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ShareFile'
        updatedAt:
          type: string
          x-stoplight:
            id: 2uikrdc3sry78
          format: date-time
        updatedBy:
          $ref: '#/components/schemas/User'
...
    Project:
      title: Project
      x-stoplight:
        id: o6laestj8lfrr
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          x-stoplight:
            id: ll4sov9aexq0v
        type:
          type: string
          enum:
            - NORMAL
            - STATION
            - OFFICE
            - TELEWORKING
          x-stoplight:
            id: 0p30i71dxed4x
        objectiveCost:
          type: number
          x-stoplight:
            id: uti97ajby18br
        startDate:
          type: string
          x-stoplight:
            id: d0ej6bz003y22
          format: date
        endDate:
          type: string
          x-stoplight:
            id: uzeirrd29qxt9
          format: date
        activityCenter:
          $ref: ../gestepm-master-data-api/api-spec.yaml#/components/schemas/ActivityCenter
        forumId:
          type: integer
          x-stoplight:
            id: dvu6fkl5xwqpw
        state:
          type: integer
          x-stoplight:
            id: tromrcn4vvmwc
        responsible:
          x-stoplight:
            id: thgh3qi3x9aoc
          type: array
          items:
            $ref: '#/components/schemas/User'
      required:
        - id
        - name
        - type
        - objectiveCost
        - startDate
        - endDate
        - activityCenter
        - state
        - responsible
    ShareFile:
      title: ShareFile
      x-stoplight:
        id: m8bximaz51t4c
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          x-stoplight:
            id: oe2w9nxlwwwkk
        content:
          type: string
          x-stoplight:
            id: prsxmvb3f9spa
        url:
          type: string
          x-stoplight:
            id: 2mfhf4p1ebytp
      required:
        - id
        - name
        - content
        - url
    SubFamily:
      title: SubFamily
      x-stoplight:
        id: mw171n2rxxhvr
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          x-stoplight:
            id: lnngd121n48ga
    Material:
      title: Material
      x-stoplight:
        id: 9erl59kl8o8oo
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
          x-stoplight:
            id: rfh6ew7z1jlea
        units:
          type: integer
          x-stoplight:
            id: u4cp15j0fz8lx
        reference:
          type: string
          x-stoplight:
            id: 7lqm8m6y0f4ih
...
    User:
      title: User
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        surnames:
          type: string
        fullName:
          type: string
          x-stoplight:
            id: enqu6ivmk02je
        email:
          type: string
          x-stoplight:
            id: 59vgomk4axlzd
        activityCenter:
          $ref: ../gestepm-master-data-api/api-spec.yaml#/components/schemas/ActivityCenter
        state:
          type: integer
          x-stoplight:
            id: 3amxnirelezop
        signingId:
          x-stoplight:
            id: vu87etbxd1yl7
          type: integer
        forumUsername:
          type: string
          x-stoplight:
            id: lzb03wvdgsp4s
        role:
          $ref: ../gestepm-app-config-api/api-spec.yaml#/components/schemas/Role
        level:
          $ref: ../gestepm-master-data-api/api-spec.yaml#/components/schemas/Level
        workingHours:
          type: number
          x-stoplight:
            id: zmfk0xio0smfd
        currentYearHolidaysCount:
          type: integer
          x-stoplight:
            id: cgd6s04xrx4po
        lastYearHolidaysCount:
          type: integer
          x-stoplight:
            id: okhm5os3bs84d
      required:
        - id
        - name
        - surnames
        - fullName
        - email
        - state
        - role
        - workingHours
        - currentYearHolidaysCount
        - lastYearHolidaysCount
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    NoProgrammedShare:
      description: Example response
      content:
        application/json:
          schema:
            type: object
            properties:
              metadata:
                $ref: ../../base-api-spec.yaml#/components/schemas/Meta
              data:
                $ref: '#/components/schemas/NoProgrammedShare'
    NoProgrammedShareList:
      description: Example response
      content:
        application/json:
          schema:
            type: object
            properties:
              metadata:
                $ref: ../../base-api-spec.yaml#/components/schemas/Meta
              data:
                type: array
                x-stoplight:
                  id: dyvkfgdppk87v
                items:
                  $ref: '#/components/schemas/NoProgrammedShare'
...
  requestBodies:
    CreateNoProgrammedShare:
      content:
        application/json:
          schema:
            type: object
            properties:
              userId:
                x-stoplight:
                  id: vwybpkwl613hp
                type: integer
              projectId:
                x-stoplight:
                  id: 2juj0v781js7f
                type: integer
            required:
              - userId
              - projectId
    UpdateNoProgrammedShare:
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              data:
                type: object
                x-stoplight:
                  id: oxud49ln29kw4
                properties:
                  userId:
                    type: integer
                  startDate:
                    type: string
                    format: date-time
                  endDate:
                    type: string
                    format: date-time
                  description:
                    type: string
                  familyId:
                    type: integer
                  subFamilyId:
                    type: integer
                  topicId:
                    type: integer
                  forumTitle:
                    type: string
                  state:
                    type: string
                    enum:
                      - NEW
                      - INITIALIZED
                      - IN_PROGRESS
                      - CLOSED
              files:
                type: array
                items:
                  type: string
                  format: binary
            required:
              - userId
...
  parameters:
    Share-Id:
      name: shareId
      in: path
      required: true
      schema:
        type: integer
      description: Share identificator.
...